requirements.txt
scikit-learn==1.2.2
flask
flask_cors
python-dotenv
werkzeug
opencv-python
tensorflow-cpu
keras
joblib
numpy<2.0.0
pandas
pillow
supabase
